pyproject.toml
[tool.poetry]
name = "culinary-django"
version = "0.1.0"
description = ""
authors = ["Evgeniy_Borovoy <borovoyevgen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.4"
python-dotenv = "^1.1.1"
pillow = "^11.3.0"
django-jazzmin = "^3.0.1"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
